+++ test/ValidatorChainTest.php
@@ -118,12 +118,6 @@ class ValidatorChainTest extends \PHPUnit_Framework_TestCase
         StaticValidator::execute('1234', 'UnknownValidator');
     }
 
-    public function testIsValidWithParameters()
-    {
-        $this->assertTrue(StaticValidator::execute(5, 'Between', [1, 10]));
-        $this->assertTrue(StaticValidator::execute(5, 'Between', ['min' => 1, 'max' => 10]));
-    }
-
     public function testSetGetMessageLengthLimitation()
     {
         AbstractValidator::setMessageLength(5);